[TEST] l10n_ve_payment_extension: corregir fallos de CI en tests de retención
CI (binauralbot) reportó 2 fallos en TESTING_INTEGRATION:

1. test_retention_lifecycle.test_02_sequence_creation_when_missing
   esperaba padding 8/5 en la secuencia recreada — ya corregido en el
   commit anterior (f3cd06249), este solo confirma que sigue en verde.

2. test_retention_sequence_no_gap.test_02_sequence_number_not_consumed_
   on_rollback fallaba con AttributeError: la clase no heredaba los
   helpers _prepare_invoice_for_retention/_create_iva_retention (solo
   existían en TestRetentionLifecycle). Se mueven ambos helpers a
   RetentionTestCommon (test_withholding_common_VEF.py) para que los
   compartan las dos clases de test sin duplicar código ni reejecutar
   la suite de TestRetentionLifecycle por herencia.

De paso se quita partner.mobile del test nuevo: ese campo no existe en
res.partner en esta versión de Odoo (se consolidó en phone).

Los otros 2 fallos reportados por CI (test_retention_payment_move_date,
fecha hardcodeada vs. fecha de ejecución) son preexistentes en
maintenance-19.0 y no fueron tocados por este PR.

References:
- Ticket: https://www.binauraldev.com/odoo/helpdesk.ticket/15000
- Tarea:
- PR: https://github.com/binaural-dev/odoo-venezuela/pull/1273
- Otros:
